// src/user/dto/user-down.dto.ts
import { IsEmail, IsString } from 'class-validator';
import { ValidateDto } from '../../types/validateDto';

export class UserDownloadDto implements ValidateDto {
	@IsEmail({}, { message: 'Неверно указан email' })
	email: string;

	jwt?: string;

	unvalidate?: string;

	path: string;
}